the client is better connected to the gui
these files wont work as actually intended until the database is attached through the correct server but has a good amount of stuff done. the logout and login functions work perfectly just need the server side to confirm that it is an actual login to handle searching through the database and stuff.